+++ client/src/components/WeatherWidget.tsx
